berlin refactored
